Update collections api to permit updating via PUT.
This implements the newer more fine grained security available in
tastypie 0.9.12. PUTs are only allowed for unlocked user owned
collections. GETs for list and detail are restricted to user owned or
public collections only, and the initial framework for DELETE is in
place but not in use due to CASCADE issues not being checked yet.